General Discord Bot upload
The initial Discord bot commit. Contains bot.py which loads all of the extensions. The extensions and bot.py use files.py to interact with the JSON files within the project. The err.log keeps track of all of the errors that occur during active state of the bot. The requirements file can be used to download all of the required Python files for the bot. The JSON file contains information that will be dumped and loaded during the active state of the bot."